+++ CHANGELOG.md
@@ -10,6 +10,10 @@
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#1834](https://github.com/realm/SwiftLint/issues/1834)
 
+* Add `switch_case_alignment` opt-in rule to `check` case and `default`  
+  statements vertically align with the enclosing `switch` statement.
+  [Austin Lu](https://github.com/austinlu)
+
 ##### Bug Fixes
 
 * Improve how `opening_brace` rule reports violations locations.